Continued separating teller and admin routes. Added comments. Moved imports only needed for one route
