+++ app/blueprints/admin/forms.py
@@ -2,9 +2,9 @@ from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField
 from wtforms.validators import Optional, Length
 
-# -----------------------------------------------------------------------------
-# AdminSettingsForm: Modify admin password only.
-# -----------------------------------------------------------------------------
+# --------------------------
+# Modify admin password only
+# --------------------------
 class AdminSettingsForm(FlaskForm):
     email = StringField('Email')
     password = PasswordField('New Password', validators=[Optional(), Length(min=8)])
